splited ValidSessionBeans.java into multiple files
