+++ src/test/resources/projects/maven/jakarta-sample/src/main/java/io/openliberty/sample/jakarta/ejb/sessionbean/ValidSingletonBeanExplicit.java
@@ -1,0 +1,10 @@
+package io.openliberty.sample.jakarta.ejb.sessionbean;
+
+import jakarta.ejb.Singleton;
+
+// Valid: public Singleton with explicit constructor
+@Singleton
+public class ValidSingletonBeanExplicit {
+    public ValidSingletonBeanExplicit() {
+    }
+}
