+++ src/test/resources/projects/maven/jakarta-sample/src/main/java/io/openliberty/sample/jakarta/ejb/sessionbean/ValidStatefulBeanNoConstructor.java
@@ -1,0 +1,8 @@
+package io.openliberty.sample.jakarta.ejb.sessionbean;
+
+import jakarta.ejb.Stateful;
+
+// Valid: public, non-final, non-abstract, top-level — no constructor (default applies)
+@Stateful
+public class ValidStatefulBeanNoConstructor {
+}
