+++ src/test/resources/projects/maven/jakarta-sample/src/main/java/io/openliberty/sample/jakarta/ejb/sessionbean/ValidStatelessBeanExplicit.java
@@ -1,0 +1,10 @@
+package io.openliberty.sample.jakarta.ejb.sessionbean;
+
+import jakarta.ejb.Stateless;
+
+// Valid: public, non-final, non-abstract, top-level — explicit no-arg constructor
+@Stateless
+public class ValidStatelessBeanExplicit {
+    public ValidStatelessBeanExplicit() {
+    }
+}
